+++ widget/gonk/nativewindow/GonkBufferQueueO/GonkBufferQueueConsumer.h
@@ -125,7 +125,7 @@ class GonkBufferQueueConsumer : public BnGonkGraphicBufferConsumer {
   // setConsumerUsageBits will turn on additional usage bits for dequeueBuffer.
   // These are merged with the bits passed to dequeueBuffer.  The values are
   // enumerated in gralloc.h, e.g. GRALLOC_USAGE_HW_RENDER; the default is 0.
-  virtual status_t setConsumerUsageBits(uint32_t usage);
+  virtual status_t setConsumerUsageBits(uint64_t usage);
 
   // setTransformHint bakes in rotation to buffers so overlays can be used.
   // The values are enumerated in window.h, e.g.
